pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "claw"
version = "0.1.0"
description = "Coding and Language Agent Workbench — sovereign AI coding agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.28.0",
    "openai>=1.35.0",
    "pgvector>=0.3.2",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.30",
    "tiktoken>=0.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "pydantic>=2.7.0",
    "pystray>=0.19.5",
    "pillow>=10.3.0",
    # video generation (GPU)
    "diffusers>=0.30.0",
    "accelerate>=0.33.0",
    "transformers>=4.44.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.5.1",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
    "watchdog>=4.0.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "pytest-asyncio>=0.23.7", "ruff>=0.4.0"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
